requirements.txt
flask
tuya-connector-python
python-dotenv
requests
openai